requirements.txt
Django==5.0.3
djangorestframework==3.14.0
web3==6.11.3
python-decouple==3.8
psycopg2-binary==2.9.9
django-ratelimit==4.1.0
pytest==8.0.0
pytest-django==4.8.0
black==24.2.0
flake8==7.0.0
django-cors-headers==4.3.1
eth-typing<4.0.0